Fix to Linux handling and ID modifications
Fixed a problem that resulted in incorrect output on Linux OS and also tidied the output up a bit by making it a function. Also added the ability to skip adding _ORF# suffixes to IDs (useful if wanting only a single ORF per sequence).

At some point it might be nice to tidy the whole program up since I've learnt a lot since I first made this (such as how to handle compatibility on Windows/Linux), but it would be a bit time intensive. Maybe later.